CameraController target can refer to entities that have not yet been added or re-added to simulation
Target no longer changes to the next entity if current target removed. Instead the camera now maintains a reference to the previous target, which is useful because the camera will remain in place at the point when the target was removed, which is the expected behavior. Also, if the target is re-added, the camera will resume tracking the same target.